Allow parsing not-quite-semver versions
Change the behaviour for parsing not-quite-semver versions, specifically
when there is only a major section, and when there is no patch section.
Rather than error out, we check if we are at the end of the string. If
so, then we return the version with all the information we know so far.
If we are not at the end of the string then we error out with the same
error message.